revise icon documentation
Screen readers would completley pass over the listing if icons as `<i class=“fa fa-*”></i>` largely means nothing to a screen reader.

Updated the example icons to have a `role=“img”` and `aria-label` to announce what type of icon was being represented.

In the `button` examples, added an `aria-hidden=“true”` attribute, so that the icons were not interpreted by screen readers.

Revised the wording of some parts of the documentation to make wording more concise.