Expand sl-tree-item on first selection.
Previously, an sl-tree with the 'single' selection mode would always toggle the expanded state of an item when selecting it, which lead to an unexpected behaviour that selecting an expanded (but not previously selected) item would collapse it.

This changes that behaviour so that an item is always (in 'single' mode) expanded if it was not previously selected. Clicking an item that was already selected continues to toggle it's expanded state.